Uploaded working code
ProCon lab now runs. The bug had to do with issues where getting the command code was including junk data because of no null-termination. Added assurance of a null-terminator. Still appears to be some issues because when running ProCon there is a large steady-state error, even with what I would expect to be good gains. Also, anti-windup is not implemented, need to find the code for that.