Fixed lab change detection
Adjusted how the SerialComms class finds a lab change because it was missing the L command before. When testing the current code, lab changes do not seem to work properly. When inputting command "L2%", a new lab is created, but this should call the default case in CUatHomeFactory, yet the normal ProCon functions still seem to work lab in the Arduino code may not be replaced like it should or something else is wrong.