New SpeedofSound; Has a bug
New SpeedofSound class that stores data and only sends on request.

Current bug: Seems to only work with small arrays because when the array size is set to something like 500 the write_index doesn't work properly for some reason. Works for 8 or so indices and then jumps to a large number and bounces back occasionally despite the modulo operation which should keep it within a small range of numbers.